Feat: Info window implmeneted.
When the user hovers on a colomn in the chart, a window pops up displaying information about represented by the bar.
jQuery hover function combined with d3 used to add and remove elements to acheive that.
This step could be done without D3, but for the sake of experimenting it was done with D3.